+++ package.json
@@ -1,10 +1,10 @@
 {
   "name": "homebridge-rachio-irrigation",
-  "version": "1.2.9",
+  "version": "1.2.10",
   "description": "Rachio Irrigation System platform plugin for [HomeBridge](https://github.com/nfarina/homebridge).",
   "main": "index.js",
   "dependencies": {
-    "axios": "^0.24.0"
+    "axios": "^0.25.0"
   },
   "engines": {
     "homebridge": "^1.4.0",
